refactor(daemon): move the Dream subdomain into src/dream/
Pure move: dream-runner.ts -> src/dream/runner.ts and memory-dreamer.ts ->
src/dream/dreamer.ts, with importer paths fixed. No symbol renames, no
decomposition, no barrel.
