+++ packages/daemon/src/cp/client.ts
@@ -140,7 +140,7 @@ import type { AgentWaker } from './agent-wake.js'
 import { AgentWakeViolationError } from './agent-wake.js'
 import type { DreamReader } from './dream-reader.js'
 import type { LocalSkillsReader } from './local-skills-reader.js'
-import { DreamViolationError, DreamStateError } from '../agents/dream-runner.js'
+import { DreamViolationError, DreamStateError } from '../dream/runner.js'
 import { ReqRep, WireError, type Clock, type TimerHandle, type Transport } from '@agentconnect.md/connection'
 import type { ConfigApply } from './config-apply.js'
 import type { Logger } from '../log.js'
